Tidy query module.
Now that query() is a method of Index it makes more sense to have the
query helper functions in the index module.

Keep the query module just for tests.

Some improvements to the handling of test utility functions and flatten
the score module hierarchy for better documentation output.
