--- src/query/score/mod.rs
@@ -1,2 +1,0 @@
-pub mod calculator;
-pub mod default;
