+++ src/score/calculator.rs
@@ -1,6 +1,5 @@
-use crate::{
-    index::{DocumentDetails, DocumentPointer, FieldDetails, InvertedIndexNode},
-    query::QueryResult,
+use crate::index::{
+    DocumentDetails, DocumentPointer, FieldDetails, InvertedIndexNode, QueryResult,
 };
 use std::{collections::HashMap, fmt::Debug};
 use typed_generational_arena::StandardIndex as ArenaIndex;
